// src/model/ui/routes/SeminarRoutesEnum.ts
export enum SeminarRoutesEnum {
    AGENDA = `/:seminarId/:seminarName/agenda`,
    ATTENDEES = '/:seminarId/:seminarName/attendees',
    SUPPORTING_MATERIALS = '/:seminarId/:seminarName/supporting-materials',
    CONTRACTS = '/:seminarId/:seminarName/contracts',
    COST_SHEET = '/:seminarId/:seminarName/cost-sheet'
}