// src/model/domain/enums/InternalPortalPermissionEnum.ts
export enum InternalPortalPermissionEnum {
    INTERNAL_SEMINAR_MANAGEMENT = "INTERNAL_SEMINAR_MANAGEMENT",
    INTERNAL_CALENDAR_MANAGEMENT = "INTERNAL_CALENDAR_MANAGEMENT",
    INTERNAL_EXAMS_MANAGEMENT = "INTERNAL_EXAMS_MANAGEMENT",
    INTERNAL_REPORTS = "INTERNAL_REPORTS",
    INTERNAL_FORMS_BUILDER = "INTERNAL_FORMS_BUILDER",
    INTERNAL_SURVEY_BUILDER = "INTERNAL_SURVEY_BUILDER",
    INTERNAL_JOURNAL_CANDIDATES = "INTERNAL_JOURNAL_CANDIDATES",
    INTERNAL_COST_SHEET = "INTERNAL_COST_SHEET",
}